backend/pyproject.toml
[tool.poetry]
name = "json-form-share-backend"
version = "0.1.0"
description = "json-form-share backend, based on Django and DJango-Rest-Framework"
authors = ["hypsug0 <fcloitre@dbchiro.org>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
mypy = "flake8"
flake8 = "^3.9.2"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
